requirements.txt
Pillow>=2.0.0
selenium==2.53.6
numpy==1.11.1
matplotlib==1.4.3
scikit-image==0.12.3
scipy==0.17.1
six==1.10.0
Cython==0.24
pytz==2016.4
